Add tests for BatchEngine
