+++ src/promptflow/tests/test_configs/eager_flows/dummy_flow_with_trace/entry.py
@@ -1,5 +1,4 @@
 import asyncio
-from dataclasses import dataclass
 
 from promptflow import trace
 
@@ -15,8 +14,7 @@ async def dummy_llm(prompt: str, model: str, wait_seconds: int):
     return prompt
 
 
-
-async def my_flow(text: str, models: list):
+async def my_flow(text: str, models: list = []):
     tasks = []
     for i, model in enumerate(models):
         tasks.append(asyncio.create_task(dummy_llm(text, model, i + 1)))
