+++ src/promptflow/tests/test_configs/eager_flows/dummy_flow_with_trace/samples.json
@@ -1,0 +1,4 @@
+[
+    {"text": "text", "models": ["model"]},
+    {"text": "text", "models": ["model", "model_2", "model_3"]}
+]
